requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
datasets==3.0.1
debugpy==1.8.6
decorator==5.1.1
dill==0.3.8
editdistance==0.8.1
epitran==1.25.1
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
g2pK==0.9.4
grpcio==1.66.2
huggingface-hub==0.25.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jamo==0.4.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
konlpy==0.6.0
langdetect==1.0.9
lingua-language-detector==2.0.2
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.3
panphon==0.21.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
polib==1.2.0
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
unicodecsv==0.14.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.13.1
